refactor: add directory for icon buttons
