+++ dev-client/src/components/SiteCard.tsx
@@ -20,8 +20,8 @@ import {StyleSheet, ViewStyle} from 'react-native';
 
 import {Site} from 'terraso-client-shared/site/siteSlice';
 
+import {IconButton} from 'terraso-mobile-client/components/buttons/icons/IconButton';
 import {Card} from 'terraso-mobile-client/components/Card';
-import {IconButton} from 'terraso-mobile-client/components/icons/IconButton';
 import {
   Box,
   Heading,
