+++ dev-client/src/components/buttons/icons/CloseButton.tsx
@@ -17,7 +17,7 @@
 
 import {PressableProps} from 'react-native-paper/lib/typescript/components/TouchableRipple/Pressable';
 
-import {IconButton} from 'terraso-mobile-client/components/icons/IconButton';
+import {IconButton} from 'terraso-mobile-client/components/buttons/icons/IconButton';
 
 type CloseButtonProps = {
   onPress?: PressableProps['onPress'];
