+++ dev-client/src/components/sheets/InfoOverlaySheetButton.tsx
@@ -15,7 +15,7 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 
-import {InfoButton} from 'terraso-mobile-client/components/buttons/InfoButton';
+import {InfoButton} from 'terraso-mobile-client/components/buttons/icons/InfoButton';
 import {InfoOverlaySheet} from 'terraso-mobile-client/components/sheets/InfoOverlaySheet';
 
 type Props = {
