+++ dev-client/src/navigation/components/AppBarIconButton.tsx
@@ -18,7 +18,7 @@
 import {
   IconButton,
   IconButtonProps,
-} from 'terraso-mobile-client/components/icons/IconButton';
+} from 'terraso-mobile-client/components/buttons/icons/IconButton';
 
 export const AppBarIconButton = (props: IconButtonProps) => (
   <IconButton size="md" _icon={{color: 'primary.contrast'}} {...props} />
