+++ dev-client/src/screens/SlopeScreen/SlopeMeterScreen.tsx
@@ -26,7 +26,7 @@ import {Button} from 'native-base';
 
 import {updateSoilData} from 'terraso-client-shared/soilId/soilIdSlice';
 
-import {BigCloseButton} from 'terraso-mobile-client/components/buttons/BigCloseButton';
+import {BigCloseButton} from 'terraso-mobile-client/components/buttons/icons/BigCloseButton';
 import {Icon} from 'terraso-mobile-client/components/icons/Icon';
 import {PermissionsRequestWrapper} from 'terraso-mobile-client/components/modals/PermissionsRequestWrapper';
 import {
