+++ dev-client/src/screens/SoilScreen/SoilScreen.tsx
@@ -35,8 +35,8 @@ import {
 } from 'terraso-client-shared/soilId/soilIdSlice';
 
 import {AddDepthModalBody} from 'terraso-mobile-client/components/AddDepthModal';
+import {IconButton} from 'terraso-mobile-client/components/buttons/icons/IconButton';
 import {Icon} from 'terraso-mobile-client/components/icons/Icon';
-import {IconButton} from 'terraso-mobile-client/components/icons/IconButton';
 import {Modal} from 'terraso-mobile-client/components/modals/Modal';
 import {
   Box,
